Flytter alt yff-relatert til egen komponent
